+++ src/components/Compose/ChannelSearchPanel.tsx
@@ -15,7 +15,7 @@ import { isSameChannel } from '@/helpers/isSameChannel.js';
 import { useCompositePost } from '@/hooks/useCompositePost.js';
 import { useIsMedium } from '@/hooks/useMediaQuery.js';
 import { useSearchChannels } from '@/hooks/useSearchChannel.js';
-import { hasRpPayload } from '@/mask/plugins/red-packet/helpers/rpPayload.js';
+import { hasRpPayload } from '@/helpers/rpPayload.js';
 import type { Channel } from '@/providers/types/SocialMedia.js';
 import { useComposeStateStore } from '@/store/useComposeStore.js';
