+++ src/components/Poll/PollButton.tsx
@@ -7,7 +7,7 @@ import { Tooltip } from '@/components/Tooltip.js';
 import { SORTED_POLL_SOURCES } from '@/constants/index.js';
 import { resolveSourceName } from '@/helpers/resolveSourceName.js';
 import { useCompositePost } from '@/hooks/useCompositePost.js';
-import { hasRpPayload } from '@/mask/plugins/red-packet/helpers/rpPayload.js';
+import { hasRpPayload } from '@/helpers/rpPayload.js';
 import { useComposeStateStore } from '@/store/useComposeStore.js';
 
 export const PollButton = memo(function PollButton() {
