+++ src/components/RedPacket/RedPacketCard.tsx
@@ -28,8 +28,8 @@ import { runInSafeAsync } from '@/helpers/runInSafe.js';
 import { useAvailableBalance } from '@/hooks/useAvailableBalance.js';
 import { useChainContext } from '@/hooks/useChainContext.js';
 import { useRefundCallback } from '@/hooks/useRefundCallback.js';
-import { HappyRedPacketV4ABI } from '@/mask/bindings/constants.js';
-import { EVMChainResolver } from '@/mask/bindings/index.js';
+import { HappyRedPacketV4ABI } from '@/mask/constants.js';
+import { EVMChainResolver } from '@/mask/index.js';
 import { useAvailabilityComputed } from '@/components/RedPacket/hooks/useAvailabilityComputed.js';
 import { useRedPacketCover } from '@/components/RedPacket/hooks/useRedPacketCover.js';
 import { ComposeModalRef } from '@/modals/controls.js';
