+++ src/components/RedPacket/hooks/useClaimCallback.ts
@@ -9,8 +9,8 @@ import { resolveRedPacketPlatformType } from '@/helpers/resolveRedPacketPlatform
 import { runInSafeAsync } from '@/helpers/runInSafe.js';
 import { waitForEthereumTransaction } from '@/helpers/waitForEthereumTransaction.js';
 import { useChainContext } from '@/hooks/useChainContext.js';
-import { HappyRedPacketV4ABI } from '@/mask/bindings/constants.js';
-import { EVMChainResolver } from '@/mask/bindings/index.js';
+import { HappyRedPacketV4ABI } from '@/mask/constants.js';
+import { EVMChainResolver } from '@/mask/index.js';
 import { useCurrentClaimProfile } from '@/components/RedPacket/hooks/useCurrentClaimProfile.js';
 import { useSignedMessage } from '@/components/RedPacket/hooks/useSignedMessage.js';
 import { FireflyRedPacket } from '@/providers/red-packet/index.js';
