+++ src/components/RedPacket/hooks/useRefundCallback.ts
@@ -3,7 +3,7 @@ import { useState } from 'react';
 import { useAsyncFn } from 'react-use';
 
 import { useChainContext } from '@/hooks/useChainContext.js';
-import { EVMWeb3 } from '@/mask/bindings/index.js';
+import { EVMWeb3 } from '@/mask/index.js';
 import { useRedPacketContract } from '@/components/RedPacket/hooks/useRedPacketContract.js';
 
 export function useRefundCallback(version: number, from: string, id?: string, expectedChainId?: ChainId) {
