+++ src/helpers/getNetworkDescriptor.ts
@@ -1,5 +1,5 @@
 import type { NetworkPluginID } from '@/constants/enum.js';
-import { getRegisteredWeb3Networks } from '@/mask/bindings/index.js';
+import { getRegisteredWeb3Networks } from '@/mask/index.js';
 
 export function getNetworkDescriptor(
     expectedPluginID: NetworkPluginID,
