+++ src/helpers/resolveExplorer.ts
@@ -1,6 +1,6 @@
 import { isValidChainId as isValidSolanaChainId } from '@masknet/web3-shared-solana';
 
-import { EVMExplorerResolver, SolanaExplorerResolver } from '@/mask/bindings/index.js';
+import { EVMExplorerResolver, SolanaExplorerResolver } from '@/mask/index.js';
 
 export function resolveAddressLink(chainId: number, address: string) {
     if (isValidSolanaChainId(chainId)) {
