+++ src/hooks/useCoinInfo.ts
@@ -1,6 +1,6 @@
 import { useQuery } from '@tanstack/react-query';
 
-import { getCoinInfo } from '@/mask/bindings/index.js';
+import { getCoinInfo } from '@/mask/index.js';
 
 export function useCoinInfo(coinId: string) {
     return useQuery({
