+++ src/mask/index.ts
