+++ src/modals/ComposeModal.tsx
@@ -38,8 +38,8 @@ import { useIsSmall } from '@/hooks/useMediaQuery.js';
 import { useSetEditorContent } from '@/hooks/useSetEditorContent.js';
 import { useSingletonModal } from '@/hooks/useSingletonModal.js';
 import type { SingletonModalRefCreator } from '@/libs/SingletonModal.js';
-import { ProfileIdentifier } from '@/mask/bindings/index.js';
-import { hasRpPayload, isRpEncrypted, updateRpEncrypted } from '@/mask/plugins/red-packet/helpers/rpPayload.js';
+import { ProfileIdentifier } from '@/mask/index.js';
+import { hasRpPayload, isRpEncrypted, updateRpEncrypted } from '@/helpers/rpPayload.js';
 import { ComposeModalRef, ConfirmModalRef } from '@/modals/controls.js';
 import { captureComposeDraftPostEvent } from '@/providers/telemetry/captureComposeEvent.js';
 import type { Channel, Post } from '@/providers/types/SocialMedia.js';
