+++ src/modals/NonFungibleCollectionSelectModal/CollectionItem.tsx
@@ -5,7 +5,7 @@ import { ChainId, SchemaType } from '@masknet/web3-shared-evm';
 import LinkIcon from '@/assets/link-square.svg';
 import { ClickableButton } from '@/components/ClickableButton.js';
 import { Image } from '@/components/Image.js';
-import { EVMExplorerResolver } from '@/mask/bindings/index.js';
+import { EVMExplorerResolver } from '@/mask/index.js';
 
 interface CollectionProps {
     collection: NonFungibleCollection<ChainId, SchemaType>;
