+++ src/providers/ethereum/ExplorerResolver.ts
@@ -1,6 +1,6 @@
 import type { ChainDescriptor } from '@masknet/web3-shared-base';
 
-import { ExplorerResolver as Resolver } from '@/mask/bindings/index.js';
+import { ExplorerResolver as Resolver } from '@/mask/index.js';
 
 class ExplorerResolver<ChainId, SchemaType, NetworkType> extends Resolver<ChainId, SchemaType, NetworkType> {
     constructor(
