+++ src/services/crossPost.ts
@@ -16,8 +16,8 @@ import { getPostFailedAt } from '@/helpers/getPostFailedAt.js';
 import { resolvePostTo } from '@/helpers/resolvePostTo.js';
 import { resolveRedPacketPlatformType } from '@/helpers/resolveRedPacketPlatformType.js';
 import { resolveSourceName } from '@/helpers/resolveSourceName.js';
-import { RedPacketMetaKey } from '@/mask/plugins/red-packet/constants.js';
-import { hasRpPayload } from '@/mask/plugins/red-packet/helpers/rpPayload.js';
+import { RedPacketMetaKey } from '@/constants/rp.js';
+import { hasRpPayload } from '@/helpers/rpPayload.js';
 import { FireflyRedPacket } from '@/providers/red-packet/index.js';
 import type { FireflyRedPacketAPI, RedPacketJSONPayload } from '@/providers/red-packet/types.js';
 import { captureComposeEvent } from '@/providers/telemetry/captureComposeEvent.js';
